requirements.txt
ray==2.0.0
ray[rllib]==2.0.0
torch
pandas
pycairo
dm_tree
Box2D
gym==0.19.0
tensorflow
wandb
matplotlib
grpcio==1.43.0
ruamel.yaml
git+https://github.com/DavidWalz/polytope-sampling.git
pypoman
scipy
mhar
